GVT-1728 Re-arrange geometry plan layout code
